Implements rudimentary video support
It is mandatory to provide a png file with the same name as the video in order to have it being previewed in Storybrew.
Osb-files including videos this way will cause the osu! parser to edit the osb (putting the video commands in the video section and deleting commands associated with them) upon certain events:
-exporting the beatmap
-entering the Design-Editor and saving the beatmap
-untested: importing the beatmap
so maps including them have to get zipped and renamed to osz manually (the same for unpacking if it is the case there as well).
Also read here for reference: https://osu.ppy.sh/forum/t/504798