.ost file specification
ost is short for osu! subtitles and seeks to provide a convenient way to setup a text file for the Lyrics-effect in Notepad using Timestamps derived from the osu! editor.

The general format works as the following:

[timestamp1] [timestamp2]
text

You can use timestamps from the Compose and the Design-Editor, so the results of the following two inputs are the same:

00:02:342 - 00:07:090 -
Hello World!

2342 7090
Hello World!

Using the Design-syntax for the first and Compose-syntax for the second timestamp (and vice versa) works as well but should be avoided for the sake of readability.

Multiple lines of Lyrics are separated by an empty line:

00:02:342 - 00:11:090 -
Hello World!

00:06:744 - 00:11:090 -
Nice weather today.

One common problem with the standard Lyrics-effect is the positioning of separate lines that are supposed to appear as a sequence and fade out as one.
When used as above the second line would obstruct the first line.

To workaround this you could do something like

00:02:342 - 00:06:744 -
Hello World!

00:06:744 - 00:11:090 -
Hello World!
Nice weather today.

This is suboptimal though as the fadeout from the first line and the fade in from the second line of "Hello World!" will be visible (especially if you decide to make the opacity configurable via code).
You would have to use separate instances of the Lyrics-effect with different y-Positions and subtitlefiles to make this work as desired.
The ost-format also allows you to display empty lines by starting a new line with the underscore character _ in order to avoid this problem.

00:02:342 - 00:11:090 -
Hello World!

00:06:744 - 00:11:090 -
_
Nice weather today.

This way the "Hello World!" font will only be obstructed by the empty line and as such be fully visible and without any undesired fade effects.
For the rare case when starting your line with an underscore is desired you can put a whitespace before and after each line of the timestamp to make sure that it is still centered with the rest of your Lyrics.

00:22:617 - 00:31:090 -
 _world = new World();